+++ packages/aisbreaker-core-nodejs/src/test-config.ts
@@ -5,3 +5,4 @@ export const OPENAI_API_KEY = process.env.OPENAI_API_KEY
 export const HUGGINGFACE_API_KEY = process.env.HUGGINGFACE_API_KEY
 export const GOOGLE_CLOUD_API_KEY_authorized_user = process.env.GOOGLE_CLOUD_API_KEY_authorized_user
 export const GOOGLE_CLOUD_API_KEY_service_account = process.env.GOOGLE_CLOUD_API_KEY_service_account
+export const REPLICATE_API_KEY = process.env.REPLICATE_API_KEY